Add files via upload
Assignment Details: 

Create a simple form with no action that makes 2 text boxes, 4 buttons and a paragraph to hold the result.  Using these controls you will implement a basic calculator.  The buttons will be the following:

Addition

Subtraction

Multiplication

Division

Each of the buttons should call a custom function you have built.  These functions will get the values out of the two text boxes and will perform the appropriate operation on them.  Once calculated, the result should be shown in the paragraph.  Make sure to take pictures of your code and pictures of of all four buttons having been clicked in your print screens.  Failure to include pictures of all button executions will result in the loss of 10 points on this activity.